Created diceRoll function and test
I created a function called diceRoll which will allow us to roll any dice type we want along with as many rolls as we want. If there is more then one roll for the dice the function will automatically add the numbers together. I also created a test for proof of functionality.